+++ CHANGELOG.md
@@ -15,7 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     inhibitness.
 - `BuilderConnector::connect_to` and `BuilderConnector::connect_with` to
   connect the builder signals to actors using `woab::Signal`.
-- `woab::route_signal` to route signals by name.
+- `woab::route_signal` to route individual signals directly from the GTK
+  object, without a builder.
 
 ### Changed
 - [**BREAKING**] Updated Actix to 0.11 and Tokio to 1.14. Consequences:
@@ -23,6 +24,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     must be started in a future (`async` block) passed the new `woab::block_on`
     function.
   - `woab::run_actix_inside_gtk_event_loop` no longer accepts a name.
+
+### Removed
+- [**BREAKING**] Removed everything related to `BuilderSignal` - the derive
+  macro, the trait, and all the builder connector methods and helper structs.
+  Use `woab::Signal` instead.
 
 ## 0.2.1 - 2021-03-18
 ### Fixed